+++ rs/src/lib.rs
@@ -16,7 +16,6 @@ pub mod parsers {
   pub mod text;
   pub mod video;
 }
-pub mod state;
 pub mod complete {
   pub(crate) mod movie;
   pub(crate) mod tag;
@@ -42,7 +41,6 @@ mod tests {
 
   use crate::complete::parse_swf;
   use crate::complete::parse_tag;
-  use crate::state::ParseState;
 
   test_expand_paths! { test_parse_movie; "../tests/movies/*/" }
   fn test_parse_movie(path: &str) {
@@ -99,9 +97,7 @@ mod tests {
       _ => 10,
     };
 
-    let mut state = ParseState::new(swf_version);
-    state.set_glyph_count(1, 11);
-    let (remaining_bytes, actual_value) = parse_tag(&input_bytes, &mut state).expect("Failed to parse");
+    let (remaining_bytes, actual_value) = parse_tag(&input_bytes, swf_version).expect("Failed to parse");
 
     let expected_path = path.join("value.json");
     let expected_file = ::std::fs::File::open(expected_path).expect("Failed to open expected value file");